fix: remove macro dependencies
- Remove proc_macro_hack
- Remove dotenv_codegen_impl
