+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Changed
 
 - Minimum Supported Rust Version is now 1.56.1
+- Removed internal `dotenv_codegen_impl` crate and `proc_macro_hack` dependency
 
 ## [0.15.6] - 2022-10-17
