+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 
-members = ["dotenv", "dotenv_codegen", "dotenv_codegen_impl"]
+members = ["dotenv", "dotenv_codegen"]
