--- dotenv_codegen_impl/Cargo.toml
@@ -1,30 +1,0 @@
-[lib]
-proc-macro = true
-
-[package]
-
-name = "dotenvy_codegen_impl"
-version = "0.15.1"
-authors = [
-  "Santiago Lapresta <santiago.lapresta@gmail.com>",
-  "Craig Hills <chills@gmail.com>",
-  "Mike Piccolo <mfpiccolo@gmail.com>",
-  "Alice Maz <alice@alicemaz.com>",
-  "Sean Griffin <sean@seantheprogrammer.com>",
-  "Allan Zhang <al@ayz.ai>",
-]
-readme = "../README.md"
-keywords = ["dotenv", "env", "environment", "settings", "config"]
-license = "MIT"
-homepage = "https://github.com/allan2/dotenvy"
-repository = "https://github.com/allan2/dotenvy"
-description = "Internal implementation for dotenvy_codegen"
-edition = "2018"
-rust-version = "1.56.1"
-
-[dependencies]
-proc-macro2 = "1"
-quote = "1"
-syn = "1"
-proc-macro-hack = "0.5"
-dotenvy = { version = "0.15", path = "../dotenv" }
