--- dotenv_codegen_impl/README.md
@@ -1,8 +1,0 @@
-# dotenvy_codgen_impl
-
-[![crates.io](https://img.shields.io/crates/v/dotenvy_codegen_impl.svg)](https://crates.io/crates/dotenvy_codgen_impl)
-[![Released API docs](https://docs.rs/dotenvy_codegen_impl/badge.svg)](https://docs.rs/dotenvy_codegen_impl)
-
-Internal implementation for [dotenvy_codgen](https://docs.rs/dotenvy_codegen_impl).
-
-This is a well-maintained fork of [dotenv_codegen_implementation](https://docs.rs/dotenv_codegen_implementation).
